# Refactor to use a single event loop for coroutines