--- a/src/CSnakes.Runtime/CPython/Coroutine.cs
+++ b/src/CSnakes.Runtime/CPython/Coroutine.cs
@@ -1,5 +1,4 @@
 using CSnakes.Runtime.Python;
-using System.Collections.Concurrent;
 
 namespace CSnakes.Runtime.CPython;
 internal unsafe partial class CPythonAPI
@@ -9,89 +8,40 @@ internal unsafe partial class CPythonAPI
         return HasAttr(p, "__await__");
     }
 
-    internal class EventLoop : IDisposable
-    {
-        private readonly PyObject? loop = null;
-
-        public EventLoop()
-        {
-            if (NewEventLoopFactory is null)
-            {
-                throw new InvalidOperationException("NewEventLoopFactory not initialized");
-            }
-            loop = NewEventLoopFactory.Call();
-        }
-
-        public bool IsDisposed { get; private set; }
-
-        private void Close()
-        {
-            if (loop is null)
-            {
-                throw new InvalidOperationException("Event loop not initialized");
-            }
-            using var close = loop.GetAttr("close");
-            close?.Call();
-        }
-
-        public void Dispose()
-        {
-            ObjectDisposedException.ThrowIf(IsDisposed, this);
-            Close();
-            loop?.Dispose();
-            IsDisposed = true;
-        }
-
-        public PyObject RunTaskUntilComplete(PyObject coroutine, CancellationToken cancellationToken)
-        {
-            if (loop is null)
-            {
-                throw new InvalidOperationException("Event loop not initialized");
-            }
-            using PyObject taskFunc = loop.GetAttr("create_task");
-            using PyObject task = taskFunc.Call(coroutine);
-            using PyObject runUntilComplete = loop.GetAttr("run_until_complete");
-
-            if (cancellationToken.CanBeCanceled)
-            {
-                // On cancellation, call task cancellation in Python
-                cancellationToken.Register(() =>
-                {
-                    using PyObject cancel = task.GetAttr("cancel");
-                    // TODO : send optional message
-                    cancel.Call();
-                });
-            }
-
-            return runUntilComplete.Call(task);
-        }
-    }
-
-    private static ConcurrentBag<EventLoop> eventLoops = [];
-    [ThreadStatic] private static EventLoop? currentEventLoop = null;
+    private static readonly Lock defaultEventLoopLock = new();
+    private static EventLoop? defaultEventLoop = null;
     private static PyObject? AsyncioModule = null;
     private static PyObject? NewEventLoopFactory = null;
 
-    internal static EventLoop GetEventLoop()
+    internal static EventLoop GetDefaultEventLoop()
     {
         if (AsyncioModule is null)
         {
             throw new InvalidOperationException("Asyncio module not initialized");
         }
-        if (currentEventLoop is null || currentEventLoop.IsDisposed)
-        {
-            currentEventLoop = new EventLoop();
-            eventLoops.Add(currentEventLoop);
-        }
-        return currentEventLoop!;
+
+        lock (defaultEventLoopLock)
+            return defaultEventLoop ??= EventLoop.RunNewForever();
     }
 
     internal static void CloseEventLoops()
     {
-        foreach (var eventLoop in eventLoops)
+        EventLoop? defaultEventLoop;
+
+        lock (defaultEventLoopLock)
         {
-            eventLoop.Dispose();
+            defaultEventLoop = CPythonAPI.defaultEventLoop;
+            CPythonAPI.defaultEventLoop = null;
         }
-        eventLoops.Clear();
+
+        if (defaultEventLoop is { } someDefaultEventLoop)
+            someDefaultEventLoop.Dispose();
     }
+
+    internal static PyObject NewEventLoop() =>
+        NewEventLoopFactory switch
+        {
+            null => throw new InvalidOperationException($"{nameof(NewEventLoopFactory)} not initialized."),
+            var some => some.Call(),
+        };
 }