--- a/src/Profile/AsyncBenchmarks.cs
+++ b/src/Profile/AsyncBenchmarks.cs
@@ -13,9 +13,26 @@ public class AsyncBenchmarks : BaseBenchmark
         mod = Env.AsyncBenchmarks();
     }
 
+    [Params(1, 10, 100, 1_000)]
+    public int N { get; set; }
+
+    [Params(0.001, 1)]
+    public double Delay { get; set; }
+
     [Benchmark]
     public async Task AsyncFunction()
     {
-        await mod!.AsyncSleepy();
+        if (N == 1)
+        {
+            await mod!.AsyncSleepy(Delay);
+        }
+        else
+        {
+            var tasks =
+                from n in Enumerable.Range(0, N)
+                select mod!.AsyncSleepy(Delay);
+
+            await Task.WhenAll(tasks);
+        }
     }
 }