--- a/src/Profile/async_benchmarks.py
+++ b/src/Profile/async_benchmarks.py
@@ -2,5 +2,5 @@ import asyncio
 from typing import Any, Coroutine
 
 
-async def async_sleepy() -> Coroutine[None, None, None]:
-    await asyncio.sleep(0.001)
+async def async_sleepy(delay: float = 0.001) -> Coroutine[None, None, None]:
+    await asyncio.sleep(delay)