## Summary

This PR refactors the Python coroutine handling in CSnakes to use a single shared event loop _and_ thread for all coroutines instead of creating a new one for each coroutine operation and per thread. The change significantly improves performance and operational behaviour (more on that later), especially when running multiple coroutines concurrently.

## Design Overview

The refactored implementation:

- uses a single, lazily-initialized, shared event loop that runs continuously.
- implements a message queue pattern for scheduling coroutines on the shared loop.
- runs the event loop on a dedicated background thread via `Task.Run`.
- supports cancellation, failure and successful completion of tasks.

## Technical Details

### Key Components

1. **Shared Event Loop**
   - A singleton `EventLoop` instance is created on demand and maintained until disposed with finalisation
   - Uses a simple lock-based lazy initialization pattern through `GetDefaultEventLoop()`
   - The [event loop runs forever](https://docs.python.org/3/library/asyncio-eventloop.html#asyncio.loop.run_forever) until it is [stopped/interrupted](https://docs.python.org/3/library/asyncio-eventloop.html#asyncio.loop.stop) by some change that's signaled via a message arriving on the queue.

2. **Message Queue**
   - A `ConcurrentQueue<Request>` is used for communication with the event loop thread.
   - Scheduling of new coroutines and cancellation of existing tasks can happen on any thread, so [`call_soon_threadsafe`](https://docs.python.org/3/library/asyncio-eventloop.html#asyncio.loop.call_soon_threadsafe) is used to safely move these operations to the thread servicing the event loop.
   - Different messages are sent to the queue to signal:
     - scheduling of a new task
     - completion of task, whether it got cancelled, failed or completed
     - stopping of the event loop (when getting disposed)

3. **Task Management**
   - `CoroutineTask` class wraps Python tasks and handles their lifecycle
   - Support for cancellation and exception handling
   - Proper disposal and resource cleanup

4. **Compatible API**
   - The public `Coroutine<TYield, TSend, TReturn>.AsTask(CancellationToken)` API remains unchanged
   - Internal implementation is significantly simplified and more robust

## Performance Improvements

At the micro-level, benchmarks demonstrate significant performance improvements, especially as number concurrent operations increase:

| Coroutine Count | Performance Improvement | Memory Overhead |
|-----------------|-------------------------|-----------------|
| 1               | 10% faster              | 764B → 1.45KB |
| 10              | 61% faster              | 6.6KB → 13.8KB |
| 100             | 69% faster              | 61KB → 134KB |
| 1000            | 76% faster              | 601KB → 1.3MB |

This is despite making more calls into Python and the _O(N<sup>2</sup>) time complexity of scanning and disposing done tasks.

While memory usage doubles, the performance improvement is substantial, especially at scale. This trade-off is well worth it for any application running multiple Python coroutines, which won't be uncommon. One must bear in mind that there is also less native memory pressure due to only a single thread and event loop being used for all coroutines. So on the balance of things, the effective and working set memory usage may not be exactly double.

The detailed benchmark is as follows:

| Diff    | Method          | N    |                 Mean | Error          |      Allocated |
| ------- | --------------- | ---- | -------------------: | -------------- | -------------: |
| Old     | `AsyncFunction` | 1    |             15.51 ms | 0.045 ms       |          764 B |
| **New** |                 |      | **14.035 ms (-10%)** | **0.4669 ms**  |    **1.45 KB** |
| Old     | `AsyncFunction` | 10   |             15.57 ms | 0.035 ms       |         6643 B |
| **New** |                 |      |  **5.998 ms (-61%)** | **0.8086 ms**  |   **13.78 KB** |
| Old     | `AsyncFunction` | 100  |             62.51 ms | 1.243 ms       |        61243 B |
| **New** |                 |      | **19.470 ms (-69%)** | **2.8330 ms**  |  **134.16 KB** |
| Old     | `AsyncFunction` | 1000 |            295.31 ms | 18.179 ms      |       601752 B |
| **New** |                 |      | **72.220 ms (-76%)** | **14.1494 ms** | **1335.27 KB** |

At the macro-level, the improvements are even more significant. Using a single even loop yields more correct and predictable operation. With the current design, increasing the sleep delay in `AsyncFunction` to 1 second and then launching 1,000 coroutines concurrently takes 47 seconds and over 20 threads (and an event loop per thread) from the thread pool to complete. With the refactored design, all thousand coroutines complete in 1.07 seconds, which is reasonable to expect! If there was a timeout associated with the overall operation of say 10 seconds, the former would always time-out and therefore never complete!

## Ideas for future/further improvements

- Adding event loop statistics and monitoring capabilities
- More efficient tracking of the task that just got done, reducing the _O(N<sup>2</sup>) time complexity
- Reduce memory usage for very large coroutine counts