@atifaziz 2025-04-29
Benchmarks from 02ff30273b87e1a34a41b3aca75030ceaa4deb73:

| Method          | N    | Delay |         Mean |      Error |     StdDev |    Gen0 |  Allocated |
| --------------- | ---- | ----- | -----------: | ---------: | ---------: | ------: | ---------: |
| `AsyncFunction` | 1    | 0.001 |    15.333 ms |  0.3052 ms |  0.4074 ms |       - |    1.56 KB |
| `AsyncFunction` | 1    | 1     | 1,008.631 ms |  1.7731 ms |  1.6586 ms |       - |    2.22 KB |
| `AsyncFunction` | 10   | 0.001 |     8.097 ms |  1.0320 ms |  3.0429 ms |       - |   14.82 KB |
| `AsyncFunction` | 10   | 1     | 1,009.712 ms |  1.7997 ms |  1.5954 ms |       - |   15.88 KB |
| `AsyncFunction` | 100  | 0.001 |    11.273 ms |  1.9347 ms |  5.5821 ms |       - |  147.45 KB |
| `AsyncFunction` | 100  | 1     | 1,006.105 ms |  6.0190 ms |  5.0261 ms |       - |  144.58 KB |
| `AsyncFunction` | 1000 | 0.001 |    47.898 ms |  0.9555 ms |  2.4147 ms | 90.9091 |  1468.6 KB |
| `AsyncFunction` | 1000 | 1     | 1,039.493 ms | 12.9844 ms | 11.5104 ms |       - | 3703.76 KB |

---

- BenchmarkDotNet v0.14.0
- Windows 11 (10.0.26100.3775)
- 13th Gen Intel Core i7-1370P, 1 CPU, 20 logical and 14 physical cores
- .NET SDK 9.0.203
  - **Host**: .NET 9.0.4 (9.0.425.16305), X64 RyuJIT AVX2
  - DefaultJob: .NET 9.0.4 (9.0.425.16305), X64 RyuJIT AVX2